Fixing some conversion issues

+++ include/optimisation/instance/load_txt.hpp
@@ -25,15 +25,15 @@ bool Load(T &optimiser, std::string const &filename) {
   };
 
   std::vector<Coupling> couplings;
-  std::unordered_map<int, int64_t> indices;
-  std::unordered_map< int, int64_t> connectivity;
+  std::unordered_map<int, uint64_t> indices;
+  std::unordered_map< int, uint64_t> connectivity;
   
-  int k = 0;
+  uint64_t k = 0;
 
   while (fin) {
     std::getline(fin, line);
 
-    int p = line.find('#');
+    uint64_t p = uint64_t(line.find('#'));
     if (p != std::string::npos) line = line.substr(0, std::size_t(p) );
     string::Trim(line);
     if (line == "") continue;
